pyproject.toml
[project]
name = "google-calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "dateparser>=1.2.2",
    "datetime>=5.5",
    "google-api-python-client>=2.183.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=6.30.1",
    "ipython>=9.6.0",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.8",
    "langgraph-swarm>=0.0.14",
    "langmem>=0.0.29",
    "langsmith>=0.4.31",
    "numpy>=2.3.3",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "slack-bolt>=1.25.0",
    "slack-sdk>=3.36.0",
]
